Add CLI convert test
